+++ app/Commands/Analyse.php
@@ -96,6 +96,7 @@ class Analyse extends Command
                 $packageAnalyser->getStepsForTable()[18],
                 $packageAnalyser->getStepsForTable()[19],
                 $packageAnalyser->getStepsForTable()[20],
+                $packageAnalyser->getStepsForTable()[21],
                 new TableSeparator,
                 [new TableCell('Ran <info>'.$amountOfAnalysisSteps.'</info> analysis steps. '.$violationText.$omitText, ['colspan' => 3])],
             ]);